+++ integration_tests/tests/conftest.py
@@ -11,7 +11,6 @@ from time import sleep
 
 import pytest
 import os
-import requests
 import zmq
 
 import flowmachine