+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added `HandsetStats` to FlowMachine.
 - Added new `ContactReferenceLocationStats` query class to FlowMachine.
 - A new zmq message `get_available_dates` was added to the flowmachine server, along with the `/available_dates`
   endpoint in flowapi and the function `get_available_dates()` in flowclient. These allow to determine the dates