+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Fixed #514, which would cause the client to hang after submitting a query that couldn't be created
+- Fixed #575, so that events at midnight are now considered to be happening on the following day
 
 ### Removed